Create BallMove.cs
The BallMove.cs script is part of a Unity project that implements the classic Pong game. This script is responsible for controlling the ball's movement, detecting collisions, and managing game state transitions related to the ball. For further explanation, check out the YouTube video: https://youtu.be/S-U8UWgl3nA